feat: root node traversal with bug fixes - Added root node interior node traversal - fixed overflow pg bug, reading filepath then reading from passed buffer - fixed pgoffset and start offset issue
Note: Root node traversal is working, checked via test db created by chatgpt which
      has interior nodes and overflow pages

+++ src/utils.rs
@@ -89,7 +89,7 @@ pub fn convert_u8_to_u16_be(buf: &[u8]) -> Box<[u16]> {
 
 pub fn read_specific_bytes(
     filepath: &String,
-    start_byte: u16,
+    start_byte: u32,
     pg_size: u16,
 ) -> Result<Vec<u8>, Box<dyn Error>> {
     let mut file = File::open(filepath)?;
